feature_store/requirements.txt
feast[redis,postgres]==0.34.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pyspark==3.0.1